add student

+++ studentmanagementsystem/src/main/java/com/studentmanagementsystem/repository/ClassRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.studentmanagementsystem.entity.Classes;
 
-public interface ClassRepository extends JpaRepository<Classes, Integer> {
+public interface ClassRepository extends JpaRepository<Classes, Long> {
 
 }
