+++ studentmanagementsystem/src/main/java/com/studentmanagementsystem/repository/MarksRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.studentmanagementsystem.entity.Marks;
 
-public interface MarksRepository extends JpaRepository<Marks, Integer> {
+public interface MarksRepository extends JpaRepository<Marks, Long> {
 
 }
