+++ studentmanagementsystem/src/main/java/com/studentmanagementsystem/repository/StudentsRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.studentmanagementsystem.entity.Students;
 
-public interface StudentsRepository extends JpaRepository<Students, Integer> {
+public interface StudentsRepository extends JpaRepository<Students, Long> {
 
 }
